backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
mysql-connector-python==8.4.0
groq==0.9.0
python-multipart==0.0.9
pydantic==2.7.4
pydantic-settings==2.3.4
python-dotenv==1.0.1
